backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
alembic
qiskit
qiskit-aer
cryptography
python-decouple
aiofiles
liboqs
